+++ src/common/CMakeLists.txt
@@ -31,6 +31,7 @@ set(efu_common_SRC
 set(efu_common_INC
   ${VERSION_INCLUDE_DIR}/common/version_num.h
   ${VERSION_INCLUDE_DIR}/common/Version.h
+  dataflow/DataObserver.h
   debug/Assert.h
   debug/Expect.h
   debug/Hexdump.h
