+++ src/common/dataflow/DataObserver.h
@@ -8,12 +8,17 @@
 
 #pragma once
 
-#include "DataEventListener.h"
+#include <cstdint>
 #include <locale>
 #include <vector>
+#include <memory>
 
+namespace Observer {
 
-namespace Timepix3 {
+template <typename DataEvent> class DataEventListener {
+public:
+  virtual void notify(const DataEvent &event) = 0;
+};
 
 template<typename DataEvent>
 class DataEventManager {
@@ -22,9 +27,21 @@ class DataEventManager {
 
     public:
         DataEventManager<DataEvent>() {};
-        void addListener(DataEventListener<DataEvent> *listener);
+        void addListener(DataEventListener<DataEvent> *listener) {
+    dataEventListeners.push_back(listener);
+}
 
-        void notifyListeners(const DataEvent &event) const;
+        void notifyListeners(const DataEvent &event) const {
+            for(const auto& listener : dataEventListeners) {
+                listener->notify(event);
+                }
+}
 };
+
+}
+
+namespace Timepix3 {
+
+using namespace Observer;
 
 }
